feat: agregar componente QuoteActions y refactorizar QuoteFooter y QuoteHeader para su integración
